Mixed changes to improve numerics and tagging (#57)
* Make tagging model initialization hyperparameter trading more elegant (the cleanup branch version works but has some dead code lying around that should be removed)

* For-now fix for backward-gradient-nans (Lion with lr=1e-3 trains stable; clamp in safe_softmax avoids 'DivBackward0 returns nan'; error_if_nonfinite in clip_grad_norm still sometimes triggered but doesnt cause harm -> turn it off)

* Add GATr jetclass training config

* update data directory

* add checkpoints for testing (could remove this later)

* accidentally changed this

For performance it would be better to set this to false, but right now I think it would be better to keep this

* Update base_experiment.py

* Hopefully fix possible bug during evaluation

* Modify batchsize in training/jc_gatr.yaml

* Update GATr configs (this is probably also already fixed in other branch; will resolve when merge)

* Fix checkpoint_every_n_steps key

* Fix jetclass evaluation fix

* Add missing checkpoint_every_n_steps key

* fix some toptaggingxl training

* fix grad_norm plotting

* update loading worker config

* remove debugging methods

---------

+++ experiments/tagging/experiment.py
@@ -319,7 +319,7 @@ class TaggingExperiment(BaseExperiment):
 
     def _batch_loss(self, batch):
         y_pred, label, tracker, _ = self._get_ypred_and_label(batch)
-        loss = self.loss(y_pred, label)
+        loss = self.loss(y_pred, label.to(torch.float32))
         assert torch.isfinite(loss).all()
 
         metrics = tracker
